update webpack with react
