--- options/options.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" href="options.css" />
-    </head>
-    <body>
-        <h4>What CSS properties would you like to search for? (check all that apply)</h4>
-        <div id="anchor">
-            <!-- <div>
-                <label class="label" htmlFor="backgroundColor">Background Color</label>
-                <input type="checkbox" id="backgroundColor" />
-            </div>
-            <div>
-                <label class="label" htmlFor="color">Color</label>
-                <input type="checkbox" id="color" />
-            </div>
-            <div>
-                <label class="label" htmlFor="fontFamily">Font Family</label>
-                <input type="checkbox" id="fontFamily" />
-            </div>
-            <div>
-                <label class="label" htmlFor="fontSize">Font Size</label>
-                <input type="checkbox" id="fontSize" />
-            </div>
-            <div>
-                <label class="label" htmlFor="fontWeight">Font Weight</label>
-                <input type="checkbox" id="fontWeight" />
-            </div>
-            <div>
-                <label class="label" htmlFor="backgroundImage">Background Image</label>
-                <input type="checkbox" id="backgroundImage" />
-            </div> -->
-        </div>
-        <div>
-            <button id="save">Save settings</button>
-        </div>
-        <!-- <div id="root">
-            <h4>What Elements would you like included? (check all tags that apply)</h4>
-            <div>
-                <label htmlFor="para">Background Color</label>
-                <input type="checkbox" id="para" />
-            </div>
-            <div>
-                <label htmlFor="color">Color</label>
-                <input type="checkbox" id="color" />
-            </div>
-            <div>
-                <label htmlFor="fontFamily">Font Family</label>
-                <input type="checkbox" id="fontFamily" />
-            </div>
-            <div>
-                <label htmlFor="fontSize">Font Size</label>
-                <input type="checkbox" id="fontSize" />
-            </div>
-            <div>
-                <label htmlFor="fontWeight">Font Weight</label>
-                <input type="checkbox" id="fontWeight" />
-            </div>
-        </div> -->
-
-        <script src="options.mjs" type="module"></script>
-    </body>
-</html>
