+++ package.json
@@ -5,7 +5,8 @@
   "main": "background.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "webpack": "webpack",
+    "start": "webpack serve --mode development --open",
+    "webpack": "webpack --mode production",
     "clean": "rimraf ./dist/",
     "build": "npm run clean && npm run webpack"
   },
@@ -23,7 +24,15 @@
   },
   "homepage": "https://github.com/asbillings07/Getter#readme",
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
+    "@fortawesome/free-regular-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "font-awesome": "^4.7.0",
     "path": "^0.12.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "webpack": "^5.6.0",
     "webpack-cli": "^4.2.0"
   },
@@ -34,10 +43,13 @@
     "@babel/preset-react": "^7.12.7",
     "@babel/runtime": "^7.12.5",
     "babel-loader": "^8.2.2",
+    "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^5.0.1",
     "html-webpack-plugin": "^4.5.0",
+    "rimraf": "^5.0.5",
     "style-loader": "^2.0.0",
     "terser": "^5.5.1",
-    "terser-webpack-plugin": "^5.0.3"
+    "terser-webpack-plugin": "^5.0.3",
+    "webpack-dev-server": "^5.0.2"
   }
 }
