--- popup/popup.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" href="popup.css" />
-        <script src="../utils/fontawesome.js"></script>
-    </head>
-    <body>
-        <div id="main"></div>
-        <i id="spinner" class="fas fa-sync-alt fa-spin"></i>
-
-        <script type="module" src="popup.mjs"></script>
-    </body>
-</html>
